64 bit precision rework

+++ obsidon/camera.odin
@@ -21,7 +21,7 @@ set_camera_position :: proc(position: Vec2) {
     renderer.set_camera_position(position)
 }
 
-set_camera_angle :: proc(angle: f32) {
+set_camera_angle :: proc(angle: f64) {
     renderer.set_camera_angle(angle)
 }
 
@@ -33,8 +33,8 @@ get_camera_position :: proc() -> Vec2 {
     return renderer.get_camera_position()
 }
 
-get_camera_angle :: proc() -> f32 {
-    return renderer.get_camera_angle()
+get_camera_angle :: proc() -> f64 {
+    return f64(renderer.get_camera_angle())
 }
 
 get_camera_zoom :: proc() -> f32 {
